refactor(ui): split filter config into separate files
Separate CompoundSearchFilter entity configs from filter chip
descriptors for clearer separation of concerns. Search filter
configs now only define input component behavior, while chip
descriptors define display rendering across reports and toolbars.
